// 3ME_UI/src/components/newOrder/homeCareProvider/homeCareSearch/searchHomeCare.enum.ts
export enum SearchHomeCareProviderModal {
  HOMECAREPROVIDER_FOUND = "homecareproviderFound",
  SEARCH_HOMECAREPROVIDER = "searchhomecareprovider",
  ADD_MANUAL_HOMECAREPROVIDER = "addmanualhomecareprovider",
  FACILITY_DATABASE_SEARCH = "facilitydatabasesearch",
  FACILITY_LIST = "facilitylist",
  FACILITY_NOTFOUND = "facilitynotfound",
  LOADER = "loader",
}